int overflow fix in LinkedBraidStructure

+++ BraidGroup/include/LinkedBraidStructure.h
@@ -188,7 +188,7 @@ class LinkedBraidStructure
   vector< BraidNode* >  backNodes;
   map< int , BraidNode > theNodes;
   
-  int maxNodeNumber;
+  uint64_t maxNodeNumber;
 };
